requirements.txt
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1
praw==7.7.1
ollama==0.1.7
beautifulsoup4==4.12.3
requests==2.32.2
